Handle different bookmarks item heights better
Bookmark items in the toolbar may not be same height on all systems,
generally the text size on Linux is bigger than on Windows for example.
This patch changes the height to use min-height on bookmark items and
removes their individual block padding. Thus, thee effectively have
variable block-padding depending on the text size.

In addition this patch removes few pre-proton workarounds because those
don't need to be supported any more.