curved-tabs: move tabs away from the window edge with pinned tabs
With positioninnedtabs, the left-most tab might end up outside of the
window under some circumstances (for example when menubar is enabled).
This patch moves tabbrowser-tabs away from the window edge equal to the
curve-size and makes the pre-tabs spacer a bit narrower to make the
space not look too large when the spacer is shown.