Make multi-row_tabs patch depend on window_cotrol support style
This change achieves multiple things:

* The sizing options for placeholder are in one place
* Improves compatibility with other styles
* Improves compatibility with other window control placements
* Makes fullscreen work
* Makes menubar not take any space when disabled

Previously, when Firefox did not have :has selector support, we needed
to create :hover triggered space to activate menubar, this took some
space from toolbar and required special styling to support translucent
toolbar background colors. But it was necessary, because otherwise there
would have been no way to access menubar items.

This still assumes that the user has menubar disabled, but that is fine,
because if menubar is enabled, then there is no reason to use this style
in the first place.