toolbars below content: urlbarview alignment fixes and pref to move tabs
When urlbar is open the panel input-container height changes based on
whether or not go-button is shown. Add a pseudo-element at the end which
has uses the same height value as go-button would.

In addition, this patch adds a pref
userchrome.toolbars-below-content.tabs-at-bottom.enabled that can be
used to make tabs as the bottom-most toolbar.