autohide bm and main toolbars: Add dpi dependent bottom margin for toolbox
On Windows 10 at least, the -1px static bottom margin we were using
causes a empty row of pixels to appear between tabs toolbar and nav-bar
when: 1. compact mode is used and 2. window scaling is at least 120% and
3. menubar is disabled.

This might be useful to be made to apply to Windows only, but that needs
testing. I suspect that this is also somewhat system dependent or
depends on more than just display scaling and density.