fake statusbar w bookmarksbar: Make Done label absolute start aligned
If bookmarks toolbar is empty (except for the flex space) the
empty-toolbar indicator element is visible, which causes the flex-space
to be be center-aligned - this is simply fixed with order: -1

In addition, when letterboxing is enabled and window is somewhat narrow,
the Done ::before pseudo can cause the statuspanel to be wider that the
<browser> element, which causes the <browser> to move when links are
being hovered. So lets just make the Done label absolute.