autohide toolbox: clamp urlbar-input box block padding
When the user exits customize mode, Firefox re-evaluates urlbar
dimensions and the result ends up with --urlbar-height becoming 0px
since the urlbar gets hidden.

The input box uses --urlbar-height to calculate its block padding which
in turn evaluates to full urlbar-height - thus making the contents of
input-box 0px leading to "invisible" input-box.

As a workaround we now clamp the padding-block calculation to 4px.