toolbars below content: use linux top margin hack to extension panels
We already apply computed top margin to notification popup and
permission popup so that they are visible. But it seems that if the
panel is close enough to screen bottom edge (such as in maximized
windows) then the panel contents fail typically fail to render. For
whatever reason, changing the default margin for
customizationui-widget-panel makes the failures much less frequent
although it doesn't appear to get rid of the issue completely.